Unwrap a facet_wrap plot

facet_unwrap(g, plot = TRUE, save = FALSE, path = ".", ...)

Arguments

g

the ggplot2 object created as normal

plot

whether to plot (or just save by turn the next args to TRUE)

save

whether to save the plot to disk

path

where to save them?

...

Arguments passed on to ggplot2::ggsave

filename

File name to create on disk.

device

Device to use. Can either be a device function (e.g. png), or one of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).

scale

Multiplicative scaling factor.

width,height,units

Plot size in units ("in", "cm", "mm", or "px"). If not supplied, uses the size of current graphics device.

dpi

Plot resolution. Also accepts a string input: "retina" (320), "print" (300), or "screen" (72). Applies only to raster output types.

limitsize

When TRUE (the default), ggsave() will not save images larger than 50x50 inches, to prevent the common error of specifying dimensions in pixels.

bg

Background colour. If NULL, uses the plot.background fill value from the plot theme.

Examples

if (FALSE) {
g <- iris %>%
  ggplot(aes(Sepal.Length)) +
  geom_histogram() +
  facet_wrap(~Species)
facet_unwrap(g)
}